requirements.txt
# Core ML and Data Processing
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Web Framework and API
flask>=3.0.0
streamlit>=1.30.0

# Database
psycopg2-binary>=2.9.0
PyYAML>=6.0.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.15.0

# ML Interpretability
# shap>=0.45.0  # Comentado temporariamente
# lime>=0.2.0   # Comentado temporariamente

# Text Processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Environment and Config
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Development Tools
# black>=23.12.0
# flake8>=6.1.0

# Production
# gunicorn>=21.2.0
# waitress>=2.1.0